Create README.md
The weather application will provide users with real-time weather information, forecasts, and other weather-related data, which can help them make better decisions about their day-to-day activities. In this Article, We Create a Weather App Using HTML, CSS, and JavaScript Codes. In this Weather app you enter the city's name or countries and it tells the weather of that area and we add 1 more function that tells there is Haze or Cloud.